pyproject.toml
[project]
name = "discordbot-autoschedule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]
